// integrador-angular/src/app/models/vatCondition.ts
export type VatCondition =
  | 'IVA_RESPONSABLE_INSCRIPTO'
  | 'IVA_RESPONSABLE_NO_INSCRIPTO'
  | 'IVA_NO_RESPONSABLE'
  | 'IVA_SUJETO_EXENTO'
  | 'CONSUMIDOR_FINAL'
  | 'RESPONSABLE_MONOTRIBUTO'
  | 'SUJETO_NO_CATEGORIZADO'
  | 'PROVEEDOR_DEL_EXTERIOR'
  | 'CLIENTE_DEL_EXTERIOR'
  | 'IVA_LIBERADO'
  | 'AGENTE_DE_PERCEPCIÓN'
  | 'PEQUEÑO_CONTRIBUYENTE_EVENTUAL'
  | 'MONOTRIBUTISTA_SOCIAL'
  | 'PEQUEÑO_CONTRIBUYENTE_EVENTUAL_SOCIAL';